chore: removed constants
