+++ lib/subscribers/core/dns/index.js
@@ -8,7 +8,29 @@ const BaseCoreSubscriber = require('../base')
 // eslint-disable-next-line n/no-unsupported-features/node-builtins
 const { tracingChannel } = require('node:diagnostics_channel')
 const shimmer = require('#agentlib/shimmer.js')
-const { INSTRUMENTED_METHODS } = require('./constants')
+const RESOLVE_METHODS = [
+  'resolve',
+  'resolve4',
+  'resolve6',
+  'resolveCname',
+  'resolveMx',
+  'resolveNaptr',
+  'resolveNs',
+  'resolvePtr',
+  'resolveSrv',
+  'resolveTxt',
+  // missing
+  // resolveAny
+  // resolveCaa
+  // resolveNaptr
+  // resolveSoa
+  // resolveTlsa
+]
+const INSTRUMENTED_METHODS = [
+  'lookup',
+  'reverse',
+  ...RESOLVE_METHODS
+]
 
 class DnsSubscriber extends BaseCoreSubscriber {
   constructor({ agent, logger }) {
